refactor: [FE] 결과스케치북 로직 일부가 훅으로 분리됨
