+++ frontend/src/hooks/useResultSketchbook.tsx
@@ -9,16 +9,21 @@ import {
     isWatchedBookState,
     maxSketchbookState,
 } from '@atoms/result';
+import { userState } from '@atoms/user';
 import useTimer from '@hooks/useTimer';
 import { GUIDE_PAGE_IDX } from '@utils/constants';
+import { emitWatchResultSketchBook, onWatchResultSketchBook } from '@game/NetworkServiceUtils';
 
 function useResultSketchbook() {
-    const isStarted = useRecoilValue(isStartedState);
+    const { isHost } = useRecoilValue(userState);
+    const [isStarted, setIsStarted] = useRecoilState(isStartedState);
+    const [isWatched, setIsWatched] = useRecoilState(isWatchedBookState);
     const isEnded = useRecoilValue(isEndedState);
-    const isWatched = useRecoilValue(isWatchedBookState);
+
     const { maxPageNum, maxBookNum } = useRecoilValue(maxSketchbookState);
-    const currentBookIdx = useRecoilValue(currentBookIdxState);
+    const [currentBookIdx, setCurrentBookIdx] = useRecoilState(currentBookIdxState);
     const [currentPageIdx, setCurrentPageIdx] = useRecoilState(currentPageIdxState);
+
     const setCanOneMoreGame = useSetRecoilState(canOneMoreGameState);
     const aSketchBookLimitTime = maxBookNum + 1;
     const interval = 2000;
@@ -26,6 +31,13 @@ function useResultSketchbook() {
         interval,
         clearTimerDeps: currentBookIdx,
     });
+
+    useEffect(() => {
+        setTimeout(() => setIsStarted(false), 3000);
+        onWatchResultSketchBook(setCurrentBookIdx, setCurrentPageIdx, setIsWatched);
+        // 가장 처음 나타나는 스케치북도 봤다고 서버에게 알린다.
+        if (isHost) emitWatchResultSketchBook(0);
+    }, []);
 
     useEffect(() => {
         if (isStarted || isWatched) return;
@@ -64,6 +76,11 @@ function useResultSketchbook() {
         setCurrentPageIdx(NextPageNumber);
     }
 
-    return { addSketchbookPage, subtractSketchbookPage };
+    function changeSketchbook(nextNum: number) {
+        const nextBookIdx = currentBookIdx + nextNum;
+        emitWatchResultSketchBook(nextBookIdx);
+    }
+
+    return { addSketchbookPage, subtractSketchbookPage, changeSketchbook };
 }
 export default useResultSketchbook;
